cache all loaded assets, even unused ones
Filled out assets.js with the entire list, imported from assets.json and switched to using indexeddb for caching the result (localstorage was too small for the jsonified version).

The cached assets are encoded with msgpack so the footprint is actually smaller than before, and the assets load faster and ?autostart is noticeably faster to get to the splash screen.

Also threw in another water effect - the wave around your head when you’re idle in the water.

All of these ROM/cache shenanigans are in prep for implementing a lot more actors. Stay tuned.