macro objects
Defined all the macro objects, and removed the need to import object IDs.

The macro objects are now in a hash, and the key is the macro name, lower cased as in the C sources. For example, “macro_goomba”, “macro_yellow_coin”. The hash is built from an array nearly identical to the C definition.

Provide alternate syntax for macro lists, similar to what I did with the other command lists.