new geo tool, some experiments
ConvertActor.rb works like the other geo and model tools, importing and translating the geo.inc and model.inc files for actors. Used it to update a few models, and will probably port it to JS when it’s settled.

In the process, updated and improved the new style syntax for gfx, vtx, etc. scripts and data structures. Nearly identical to C.

Also experimenting with some alternative looks for the imports and exports. Renaming some constants, like LEVEL_CASTLE_GROUNDS to CASTLE_GROUNDS and importing as LEVEL, so that it’s LEVEL.CASTLE_GROUNDS. For example.

Oh yeah, imported and commented out the rest of the global_scripts.