Solid bobombs again (GEO_BILLBOARD fix)
Fixed mishandling of args in GEO_BILLBOARD when using the alternative style of layout. The param ‘args’ is always present so can’t be tested for nulll-ness.

Reworked once more the implementation of the alternate style. This is converging on something good. I liked the string names style because it allowed lazy linking later. Turns out that isn’t necessary if the alternate methods are defined after the leyout engine is instantiated.

Note also that we’re back to command/args objects. If we go with this style and convert the intermediate alternate styles to it, the layout parser can be reverted back to its original form, without needing to lazy load.