Remove string-style geo layouts.
Moving to final two styles - the JS native original style and the alternative style that creates the same layouts in memory but does it with wrapper functions that look like the C decomp.

I can pull the switch at any time and convert the other actors to the new style.