Joe's Actors major update, 3 parts and several commits
Actors Part 1 - the new ones

Included all the textures in assets.js, and added them to the rom texture loader. This got big, so I changed over from LocalStorage to IndexedDB.

Used the Convert tool to create the missing actors’ geo, model, collision, and anims files. So that the textures have some place to go.

(Sorry about writing a new tool that overlaps the js tools; this was a learning experience.)

Actors Part 2 - Intermission

Some misc textures and code changes for actors.

Support for the alternate style for geo layouts.

Some new gbi constants and functions, some of which needs to be double-checked.

A complete set of model_ids, object_constants and surface_terrains.

A new programmatic way to make skyboxes.

Actors Part 3 - updated ones

Update a bunch of existing actors to the new layout style. Use the Convert tool to bring em in, and finish a few of them by hand.