Actors Part 1 - the new ones
Included all the textures in assets.js, and added them to the rom texture loader. This got big, so I changed over from LocalStorage to IndexedDB.

Used the Convert tool to create the missing actors’ geo, model, collision, and anims files. So that the textures have some place to go.

(Sorry about writing a new tool that overlaps the js tools; this was a learning experience.)