finish fix for inconsistent node arg
From discord:
I fixed a bug in my last PR with the way that geo_ routines are handled, the ones that are called with GEO_ASM as parts of models. The routines are called in different phases, like GEO_CONTEXT_CREATE when initing the level, GEO_CONTEXT_RENDER to actually render, and so on. They're called with a node arg, and the bug was that the node arg was inconsistent, for some phases it was a graphnode and for others it was the wrapper. This is a JSism and doesn't exist in the C code. I fixed it to be consistent, but didn't fix all the places that depended on the broken behavior.

(Also undid my DEBUG that skips the Mario Head scene when launching.)