Snowy Glider Jungle and website fixes (#614)
* Lots of site rewrite fixes

Updated a lot of lines to make the site more responsive, put the chat aside the game screen, and lots more other improvements.

* Customizer menu hotfix

Changed the background image to the correct one. Nice job Eros you already messed something up :trollface:

* Some important problems solved, buttons changed

Some stuff has been fixed since the site didn't work properly yet, still not perfect.
New problems have been commented on by me so make sure to check those out if possible.

* Game canvas and chat size hotfixes

* Hotfix #2

* Hotfix #3

* Make the chat not disappear when in widescreen mode

* Disable the respawn button's update code

* Fix seesaw flag bug

The flag's model is replaced with the seesaw likely because of model
loading in the script being different due to the new level commands.

* Last fixes for a while

As requested, the boxes over the gameCanvas are all the same size now.
The controls panel has been reorganized.
And the chat has been fixed in widescreen mode.

* Revert respawn button to old css method

This will make it look a bit worse but it'll work again, I didn't want to leave it not functioning as intended.

* Update Mario and Luigi cap textures

* Remove leveling system from display name

* Fix site bugs

* Minor size adjustment

* Save widescreen setting

Untested but should work in theory.

* Remove useless coment

* Snow in GJ

* Stage index.js for snow

* Finish up rest of snow system

* Snowy Glider Jungle

Also decided to just render snow on textCanvas.

* Fix viewport bug

* HTML fixes

* Update caniuse-lite

* Revert caniuse-lite update, code breaker

* Fix rules z-index

Chatlog was overlaying it.

* Remove need of fxCanvas

* Fix music div placement

Co-authored-by: eros71-dev <16540103+eros71-dev@users.noreply.github.com>
Co-authored-by: Agent-11 / Agent X <Agent-11@users.noreply.github.com>